// src/app/models/logs.model.ts
export class Logs {
    id?: any;
    nombreLog?: string;
    descripcionLog?: string;
    tipoLog?: string;
    fueGenLog?: string;
    sincronizacionLog?: string;
    formatoLog?: string;
    retencionLog?: string;
    backupLog?: string;
    freBacLog?: string;
    monitoreoLog?: string;
    idSer?: any;
    idApl?: any;
    
}